news_board/serializers.py
from rest_framework import serializers
from .models import Article, Comment


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = (
            "id",
            "title",
            "creation_date",
            "amount_of_upvotes",
            "author_name",
            "comments",
        )
        read_only_fields = ["comments"]


class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = "__all__"
