requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b0
click==7.1.2
Django==3.2.2
django-crontab==0.7.1
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
